Add jsonschema-circe and allow us to share code better
