+++ jsonschema-circe/src/main/scala/sttp/apispec/circe.scala
@@ -1,0 +1,8 @@
+package sttp.apispec
+
+import sttp.apispec.AnySchema.Encoding.Boolean
+import sttp.apispec.internal.{JsonSchemaCirceDecoders, JsonSchemaCirceEncoders}
+
+object circe extends JsonSchemaCirceEncoders with JsonSchemaCirceDecoders {
+  override def anyObjectEncoding: AnySchema.Encoding = Boolean
+}
